compiletest: add `{ignore,only}-riscv32cheriot-unknown-cheriotrtos` directive (#61)
This is useful to disable/enable tests that need CHERIoT's `core` being
compiled when the target is another platform (otherwise `core` for
CHERIoT may not be available yet and the test would fail).
It is also useful when other directives could match CHERIoT, but we
actually want to ignore that test.